An application that uses JSON feeds, and parses them, to form intelligent data. It will display the badges, points and relevant data about Kongregate badges.
It dynamically loads 95% or so of its content, so it would be a great place to look at loading images dynamically, loading JSON files too.

It parses the user’s name through the file name (after the underscore), but you can change the username by mousing over the username in the top left. But only after the user’s badges have finished loading.